added check if zookeeper is built

+++ tests/test_zookeeper.sh
@@ -47,7 +47,7 @@ zookeeper="zookeeper-$ZOOKEEPER_VERSION"
 
 cd "$srcdir/..";
 
-if [ -z "$(netstat -an | grep [:.]2181)" ]; then
+if [ -n "$zookeeper_built" -a -z "$(netstat -an | grep [:.]2181)" ]; then
     cp -vf "$zookeeper/conf/zoo_sample.cfg" "$zookeeper/conf/zoo.cfg"
 
     "$zookeeper/bin/zkServer.sh" start &
@@ -60,9 +60,11 @@ $perl -T $I_lib ./check_zookeeper.pl -s -w 10 -c 20 -v
 hr
 $perl -T $I_lib ./check_zookeeper_config.pl -C "$zookeeper/conf/zoo.cfg" -v
 hr
-$perl -T $I_lib ./check_zookeeper_child_znodes.pl -z / --no-ephemeral-check -v
-hr
-$perl -T $I_lib ./check_zookeeper_znode.pl -z / -v -n --child-znodes
-hr
+if [ -n "$zookeeper_built" ]; then
+    $perl -T $I_lib ./check_zookeeper_child_znodes.pl -z / --no-ephemeral-check -v
+    hr
+    $perl -T $I_lib ./check_zookeeper_znode.pl -z / -v -n --child-znodes
+    hr
+fi
 
 echo; echo
